Serve static document from JAR file or resources dir
Whether one or the other is done is determined at compile time (see
HTTPServer.java). Serving resources from resources directory is only
convenient during development since it allows to change documents
without recompiling the project but the PWD must be the root of the
source directory for this to work (which is inconvenient when deploying
the app).